app/requirements.txt
fastapi
uvicorn
huggingface_hub==0.22.2
pinecone-client
python-dotenv
FlagEmbedding
peft